SDR-317 refactor: UserRepository를 repository 패키지로 이동
