+++ be/src/main/java/com/jjikmuk/sikdorak/user/user/repository/UserRepository.java
@@ -1,5 +1,6 @@
-package com.jjikmuk.sikdorak.user.user.domain;
+package com.jjikmuk.sikdorak.user.user.repository;
 
+import com.jjikmuk.sikdorak.user.user.domain.User;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
